+++ src/main/java/ru/game/kiker/service/ScoreService.java
@@ -5,6 +5,6 @@ import ru.game.kiker.model.entity.OnlineGame;
 public interface ScoreService {
     OnlineGame findOnlineGameById(Long id);
 
-    Boolean updateScore(Long id, Long firstTeam, Long secondTeam, String type);
+    Boolean updateScore(Long id, Integer firstTeam, Integer secondTeam, String type);
 
 }
